Rename database.py
